pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "picky"
version = "1.0.0"
description = "Picky - Meal planning application with Flask backend"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "flask==3.0.3",
    "flask-cors==4.0.1",
    "python-dotenv==1.0.0",
    "azure-cosmos==4.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "tests*", "scripts*"]

